### 2025-02-21 #7833 @editorialbot
[PRE REVIEW]: PyPaperRetriever: A Python Tool for Finding and Downloading Scientific Literature
**Submitting author:** @JosephIsaacTurner (<a href="http://orcid.org/0009-0009-2423-3451">Joseph Turner</a>)
**Repository:** https://github.com/JosephIsaacTurner/pypaperretriever/
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
