### 2025-04-15 #7833 @arfon
@sara-shiho @SamHames @sigven @xavieryao @xavieryao :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _PyPaperRetriever: A Python Tool for Finding and Downloading Scientific Literature_

Many thanks
Arfon
