### 2025-04-18 #7833 @arfon
@musabgultekin @bagheria @reality 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is _PyPaperRetriever: A Python Tool for Finding and Downloading Scientific Literature_

Many thanks
Arfon
