### 2025-04-29 #7833 @arfon
@timClicks @maladeep @PetrKorab 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is PyPaperRetriever: A Python Tool for Finding and Downloading Scientific Literature

Many thanks
Arfon
