### 2025-05-22 #8135 @PetrKorab
@arfon Dear Editor,
             This software implements Sci-Hub as part of the API. This platform has faced some controversies regarding copyright. See e.g. [here](https://www.ifis.org/en/research-skills-blog/the-dangers-of-sci-hub). To completely follow the research integrity principles, I'd like to ask you to clarify if research software using Sci-Hub is in line with the JOSS guidelines. I can consequently proceed with the review. Thanks.
