### 2025-05-29 #8135 @xavieryao
@arfon and @JosephIsaacTurner

I have completed my review.

The described software is similar to https://github.com/ferru97/PyPaperBot but can check if a DOI exists on Unpaywall, PMC or Crossref (and SciHub if needed) and download the full text.

However, Unpaywall itself already checks if a DOI has full text in both PMC and Crossref.  See [Unpaywall sources](https://unpaywall.org/sources). It also has [integrations](https://unpaywall.org/integrations) for CLI tools, exporting citations, etc. that PyPaperRetriever implements.

Therefore, I recommend a major revision to address the need for this software and whether it meets "substantial scholarly effort".

That being said, the software package and paper themselves are well-built. The package follows all guidelines for well-structured high-quality open source software. The only thing missing is a guideline for contributors.
