### 2025-07-14 #8135 @arfon
Hi @PetrKorab – Thanks for flagging this. Your point about Sci-Hub’s copyright status is well-taken, and we appreciate you checking that the submission aligns with JOSS’s research-integrity principles. While we do not have policies that _directly_ map here, I do think there are some issues with the current approach.

@JosephIsaacTurner – Following up on the reviewer’s point: my read of the code is that the `allow_scihub` option is currently True by default. Given possible ethical and legal issues with this, I think this should be set by default to `False`. So please:

1. Change the default value of allow_scihub (or its equivalent) to False.
2. Update any documentation and examples to show Sci-Hub access as an opt-in feature.
3. Add a brief disclaimer in the README noting that users are responsible for ensuring they have the legal right to download any content.

Once those updates are in place, I'd be happy to proceed with the review.
