### 2025-07-20 #8135 @PetrKorab
@JosephIsaacTurner The project deserves good documentation. Currently, there is an examples notebook and the README. Please create a docs page where all features will be explained, including methods, parameters, etc. This will expand the package's user base and make the use more straightforward. Thanks!
