### 2025-08-09 7de922d
Add MkDocs documentation and deployment workflow
