+++ docs/installation.md
@@ -1,0 +1,13 @@
+# Installation
+
+Install **PyPaperRetriever** with pip:
+
+```bash
+pip install pypaperretriever
+```
+
+To work on the documentation locally:
+
+```bash
+pip install mkdocs mkdocs-material mkdocstrings[python]
+```
