+++ docs/index.md
@@ -1,0 +1,10 @@
+# PyPaperRetriever
+
+Welcome to the official documentation for **PyPaperRetriever**, a Python library for retrieving and processing academic papers.
+
+Use the navigation on the left to explore the docs or jump to a section below:
+
+- [Installation](installation.md)
+- [Quickstart](quickstart.md)
+- [API Reference](api.md)
+- [FAQ](faq.md)
