### 2025-08-09 77995cb
Add return type hints for documentation
