### 2025-08-09 8dbebe1
set allow_scihub default to False
