### 2025-08-10 14bc669
update paper text
