+++ paper/paper.md
@@ -9,21 +9,23 @@ tags:
 authors:
   - name: Joseph I Turner
     orcid: 0009-0009-2423-3451
-    affiliation: 1
+    affiliation: [1, 2]
   - name: Kaydance D Turner
-    affiliation: 2
+    affiliation: 3
 affiliations:
- - name: Center for Brain Circuit Therapeutics, Harvard Medical School
+ - name: NYU Grossman School of Medicine
    index: 1
- - name: Department of Computer Science, Brigham Young University
+ - name: Center for Brain Circuit Therapeutics, Harvard Medical School
    index: 2
+ - name: Department of Computer Science, Brigham Young University
+   index: 3
 date: 29 January 2025
 bibliography: paper.bib
 
 ---
 # Summary
 
-PyPaperRetriever is a Python tool designed to simplify the process of finding and downloading scientific literature. For any given DOI or PubMed ID, PyPaperRetriever communicates with multiple APIs, including Unpaywall, NIH's Entrez, CrossRef, and Sci-Hub, to locate and download PDFs directly to the user's system. The tool is user-friendly, featuring a command-line interface and functionality that can be imported into Python scripts for integration with research workflows. PyPaperRetriever offers a significant advantage over other tools by leveraging multiple APIs to ensure broad coverage, prioritizing open-access sources, and enabling PubMed ID-based searches. Additionally, PyPaperRetriever allows users to programmatically search PubMed with custom queries and download PDFs of search results, making it an essential resource for researchers conducting literature reviews or managing large datasets.
+PyPaperRetriever is a Python tool that automates the discovery and retrieval of scientific literature, starting from either a DOI or PubMed ID. It queries multiple sources—Unpaywall, NIH Entrez/PMC, CrossRef, and optionally Sci-Hub—to locate and download PDFs, prioritizing open access. The tool supports programmatic PubMed searches with custom queries, enabling bulk PDF downloads, metadata resolution, and optional figure extraction in a single, reproducible workflow. Usable via command line or as a Python module, PyPaperRetriever provides a scalable, PubMed-first solution for literature reviews, dataset creation, and other projects requiring comprehensive literature aggregation.
 
 ### Core Features
 
@@ -40,9 +42,18 @@ PyPaperRetriever is a Python tool designed to simplify the process of finding an
 **4. Finding references, and building citation networks:** PyPaperRetriever allows users to track a paper's citation network using a DOI or PubMedID. It finds papers that reference the given paper, as well as the papers it references. This process can be repeated recursively to any desired depth, enabling users to explore how the paper has influenced others and the foundational work it is built upon across multiple generations of citations.
 
 # Statement of Need
-Efficient access to scientific literature is critical for researchers, yet manually locating and downloading PDFs is often tedious and time-consuming [@Zhang:2023; @Singh:2011]. PyPaperRetriever automates this process, enabling researchers to quickly find and download papers from a variety of sources. The tool has been extensively used in research projects, including aggregating over 7,000 brain lesion case reports for LesionBank.org and supporting several thousand additional downloads for other miscellaneous projects studying brain lesions. By streamlining literature retrieval and integrating advanced search capabilities, PyPaperRetriever addresses a need in the research community, offering a practical, time-saving solution for projects requiring comprehensive literature aggregation. This workflow is easily integrated into CADIMA pipelines, where it has been successfully used to streamline data collection and analysis.
 
[…]
