+++ docs/api/paper_retriever.md
@@ -1,0 +1,17 @@
+# `paper_retriever`
+
+> Download a single paper by DOI/PMID.
+
+::: pypaperretriever.paper_retriever.PaperRetriever
+    options:
+      members:
+        - __init__
+        - download
+        - check_open_access
+        - check_pubmed_central_access
+        - check_crossref_access
+        - check_scihub_access
+      inherited_members: false
+
+## CLI entry point
+::: pypaperretriever.paper_retriever.main
