+++ docs/api/paper_tracker.md
@@ -1,0 +1,5 @@
+# `paper_tracker`
+
+> Trace a seed paper’s citation network.
+
+::: pypaperretriever.paper_tracker.PaperTracker
