+++ docs/quickstart.md
@@ -18,10 +18,13 @@ from pypaperretriever import PaperRetriever
 retriever = PaperRetriever(
     email="your.email@gmail.com",
     doi="10.7759/cureus.76081",
-    download_directory="PDFs"
+    download_directory="PDFs",
+    allow_scihub=True # If False, will only use oepn access sources
 )
 retriever.download()
 ```
+
+If `allow_scihub=True`, PyPaperRetriever will fall back to Sci-Hub if no open access copy is found. Sci-Hub is not used by default due to legal and ethical considerations.
 
 To download via PubMed ID:
 
@@ -31,9 +34,16 @@ from pypaperretriever import PaperRetriever
 retriever = PaperRetriever(
     email="your.email@gmail.com",
     pmid="33813262",
-    download_directory="PDFs"
+    download_directory="PDFs",
+    allow_scihub=True
 )
 retriever.download()
+```
+
+Or by CLI:
+
+```bash
+pypaperretriever --doi 10.7759/cureus.76081 --email your.email@gmail.com --dwn-dir PDFs --allow-scihub true
 ```
 
 ## Search PubMed and download results
@@ -45,7 +55,7 @@ search_query = "brain lesion case reports"
 searcher = PubMedSearcher(search_string=search_query, email="your.email@gmail.com")
 
 results = searcher.search(count=5)
-searcher.download_articles(download_directory="PDFs")
+searcher.download_articles(download_directory="PDFs", allow_scihub=True)
 ```
 
 ## Extract images from PDFs
@@ -70,3 +80,7 @@ tracker = PaperTracker(
 )
 results = tracker.track_paper()
 ```
+
+## What's next?
+Explore the full API reference for more details on available methods and options:
+- [API Reference](api/index.md)
