+++ docs/api/reference_retriever.md
@@ -1,0 +1,5 @@
+# `reference_retriever`
+
+> References and cited-by utilities.
+
+::: pypaperretriever.reference_retriever.ReferenceRetriever
