+++ docs/api/cli.md
@@ -1,0 +1,9 @@
+# Command-line Interface
+
+
+## Download by DOI
+Download a PDF by DOI:
+
+```bash
+pypaperretriever --doi 10.7759/cureus.76081 --email you@gmail.com --dwn-dir PDFs --allow-scihub true
+```
