+++ docs/api/pubmed_searcher.md
@@ -1,0 +1,19 @@
+# `pubmed_searcher`
+
+> Search PubMed and run a batch pipeline (download, images, refs).
+
+::: pypaperretriever.pubmed_searcher.PubMedSearcher
+    options:
+      members:
+        - __init__
+        - search
+        - download_articles
+        - extract_images
+        - fetch_references
+        - fetch_cited_by
+        - fetch_abstracts
+        - get_abstract
+        - download_xml_fulltext
+        - save
+        - save_abstracts_as_csv
+      inherited_members: false
