+++ docs/api/utils.md
@@ -1,0 +1,8 @@
+# `utils`
+
+> Helpers exposed for advanced workflows.
+
+::: pypaperretriever.utils
+    options:
+      filters:
+        - "!^_"
