+++ docs/api/image_extractor.md
@@ -1,0 +1,5 @@
+# `image_extractor`
+
+> Pull bitmap images out of PDFs.
+
+::: pypaperretriever.image_extractor.ImageExtractor
