+++ mkdocs.yml
@@ -37,6 +37,9 @@ plugins:
           options:
             docstring_style: google
             show_source: true
+            source_repository: "https://github.com/josephisaacturner/pypaperretriever"
+            source_branch: "main"
+            source_directory: "pypaperretriever"
             show_root_heading: true
             show_object_full_path: false
             separate_signature: true
