### 2025-08-10 9adcb61
docs: improve package docstrings
