+++ pypaperretriever/utils.py
@@ -1,9 +1,16 @@
 from Bio import Entrez
 import requests
 
+
 def entrez_efetch(email, id):
-    """
-    Fetches a record from the Entrez API.
+    """Fetch a record from the Entrez API.
+
+    Args:
+        email (str): Email address required by the Entrez API.
+        id (str): PubMed identifier to fetch.
+
+    Returns:
+        Any: Parsed record returned by the Entrez API.
     """
     Entrez.email = email
     handle = Entrez.efetch(db="pubmed", id=id, retmode="xml")
@@ -11,9 +18,19 @@ def entrez_efetch(email, id):
     handle.close()
     return records
 
+
 def pmid_to_doi(pmid, email):
-    """
-    Converts a PMID to a DOI using the Entrez API.
+    """Convert a PMID to a DOI using the Entrez API.
+
+    Args:
+        pmid (str): PubMed identifier to convert.
+        email (str): Email address required by the Entrez API.
+
+    Returns:
+        str | None: DOI if found, otherwise ``None``.
+
+    Raises:
+        ValueError: If no DOI is associated with the PMID.
     """
     records = entrez_efetch(email, pmid)
     try:
@@ -22,13 +39,23 @@ def pmid_to_doi(pmid, email):
             if element.attributes.get('IdType') == 'doi':
                 return str(element)
         raise ValueError("No DOI found for this PMID")
-    except Exception as e:
+    except Exception as e:  # pragma: no cover - informative print
         print(f"Error processing PMID {pmid}: {e}")
     return None
 
+
 def doi_to_pmid(doi, email):
-    """
-    Converts a DOI to a PMID using the Entrez API, and if that fails, uses the PMC API.
+    """Convert a DOI to a PMID.
+
+    The function first queries the Entrez API. If that fails, the PMC ID converter
+    service is used as a fallback.
+
+    Args:
+        doi (str): Digital Object Identifier to convert.
+        email (str): Email address required by the Entrez API.
+
+    Returns:
+        str | None: PMID if found, otherwise ``None``.
     """
     Entrez.email = email
     try:
@@ -38,7 +65,7 @@ def doi_to_pmid(doi, email):
         if search_results['IdList']:
             pmid = search_results['IdList'][0]
             return pmid
-    except Exception as e:
+    except Exception as e:  # pragma: no cover - informative print
         print(f"[ReferenceRetriever] Error converting DOI {doi} to PMID via Entrez: {e}")
 
     try:
@@ -50,19 +77,37 @@ def doi_to_pmid(doi, email):
             pmid = data.get("records", [{}])[0].get("pmid", None)
             if pmid:
                 return pmid
-        else:
-            print(f"[PyPaperRetriever] PMC ID Converter request failed with status code: {response.status_code}")
-    except Exception as e:
+        else:  # pragma: no cover - network behaviour
+            print(
+                f"[PyPaperRetriever] PMC ID Converter request failed with status code: {response.status_code}"
+            )
+    except Exception as e:  # pragma: no cover - informative print
         print(f"[PyPaperRetriever] Error converting DOI {doi} to PMID via PMC ID Converter: {e}")
     return None
 
+
 def encode_doi(doi):
-    """Encodes a DOI for safe inclusion in a filename using URL encoding."""
+    """Encode a DOI for safe inclusion in file names.
+
+    Args:
+        doi (str): DOI to encode. It may include a ``doi.org`` prefix.
+
+    Returns:
+        str: URL-encoded DOI suitable for use in file paths.
+    """
     doi = doi.split("doi.org/")[-1].split("?")[0]
[…]
