+++ pypaperretriever/image_extractor.py
@@ -1,3 +1,5 @@
+"""Utilities for extracting figures from PDF documents."""
+
 import os
 import fitz as pymupdf
 import numpy as np
@@ -9,23 +11,30 @@ import io
 
 
 class ImageExtractor:
-    """Class to extract figures from PDFs. 
-    Able to handle both native and image-based PDFs.
+    """Extract figures from a PDF file.
+
+    The extractor handles both native PDFs (containing embedded images) and
+    scanned PDFs where each page is an image.
+
+    Args:
+        pdf_file_path (str): Path to the PDF file to process.
 
     Attributes:
         filepath (str): Path to the PDF file.
-        dir (str): Directory of the PDF file.
-        is_native_pdf (bool): Whether the PDF is native or image-based.
-        img_paths (list): List of paths to extracted images.
-        img_counter (int): Shared counter for image numbering.
-    
-    Methods:
-        extract_images(): Wrapper function to extract images based on PDF type.
-        extract_from_native_pdf(): Extract figures from a native PDF using PyMuPDF.
-        handle_image_based_pdf(): Convert image-based PDF pages to images and extract figures.
+        dir (str): Directory containing the PDF file.
+        is_valid_pdf (bool): Whether the file can be opened by PyMuPDF.
+        is_native_pdf (bool): ``True`` if the PDF contains embedded text.
+        img_paths (list[str]): Paths to extracted image files.
+        img_counter (int): Counter used to name extracted images.
+        id (str | None): Optional identifier prefix for saved images.
     """
+
     def __init__(self, pdf_file_path):
-        """Init method for ImageExtractor."""
+        """Initialize the extractor.
+
+        Args:
+            pdf_file_path (str): Path to the PDF file to process.
+        """
         self.filepath = pdf_file_path
         self.dir = os.path.dirname(pdf_file_path)
         self.is_valid_pdf = False
@@ -39,7 +48,14 @@ class ImageExtractor:
             self._get_metadata()
 
     def extract_images(self):
-        """Wrapper function to extract images based on PDF type."""
+        """Extract images from the PDF.
+
+        The method determines the PDF type and delegates to the appropriate
+        extraction routine. Extracted image paths are stored in ``img_paths``.
+
+        Returns:
+            ImageExtractor: The current instance with ``img_paths`` populated.
+        """
         if not self.is_valid_pdf:
             print("PDF is not valid.")
             return self
@@ -50,7 +66,10 @@ class ImageExtractor:
             self.handle_image_based_pdf()
 
     def extract_from_native_pdf(self):
-        """Extract figures from a native PDF using PyMuPDF."""
+        """Extract figures from a native PDF using PyMuPDF.
+
+        Saves each valid image to disk and records its file path.
+        """
         try:
             with pymupdf.open(self.filepath) as doc:
                 for page_num in range(len(doc)):
@@ -109,7 +128,11 @@ class ImageExtractor:
             print(f"Error extracting from native PDF: {e}")
 
     def handle_image_based_pdf(self):
-        """Convert image-based PDF pages to images and extract figures."""
+        """Process a scanned PDF.
+
+        Each page is converted to an image and potential figures are extracted
+        using :meth:`_crop_boxes_in_image`.
+        """
         try:
             pages = convert_from_path(self.filepath, 300)  # DPI set to 300 for good quality
 
@@ -124,7 +147,14 @@ class ImageExtractor:
             print(f"Error handling image-based PDF: {e}")
 
     def _crop_boxes_in_image(self, img_path):
-        """Crop boxes (figures) from the given image and save them, filtering out small areas."""
+        """Crop figure-like regions from an image.
+
+        Args:
[…]

[…]
