### 2025-08-10 1b45592
chore: add type hints and doc build deps
