+++ pypaperretriever/paper_retriever.py
@@ -1,46 +1,47 @@
-import os
-import requests
-import time
-import random
-from bs4 import BeautifulSoup
-import json
-from glob import glob
-import fitz as pymupdf
+"""Retrieve and download scientific papers from various sources."""
+
+from __future__ import annotations
+
 import argparse
+import json
+import os
+import random
 import re
+import time
+from glob import glob
 from urllib.parse import urljoin, urlparse
 
-from .utils import entrez_efetch, pmid_to_doi, doi_to_pmid, encode_doi, decode_doi
+import fitz as pymupdf
+import requests
+from bs4 import BeautifulSoup
+from typing import Self
+
+from .utils import decode_doi, doi_to_pmid, encode_doi, entrez_efetch, pmid_to_doi
+
 
 class PaperRetriever:
-    """
-    A class to find and download scientific papers.
+    """Find and download scientific papers.
+
+    The class queries several services (Unpaywall, PubMed Central, Crossref and
+    optionally Sci-Hub) to locate a PDF for a given DOI or PMID.
+
+    Args:
+        email (str): Email address used for API requests.
+        doi (str, optional): Digital Object Identifier of the paper.
+        pmid (str, optional): PubMed identifier of the paper.
+        allow_scihub (bool, optional): Whether to query Sci-Hub as a fallback.
+        download_directory (str, optional): Directory where PDFs are stored.
+        filename (str, optional): Custom filename for the downloaded PDF.
+        override_previous_attempt (bool, optional): Overwrite existing downloads.
+
     Attributes:
-        email (str): The email address used for API requests.
-        doi (str): The DOI of the paper.
-        pmid (str): The PubMed ID of the paper.
-        allow_scihub (bool): Whether to allow searching Sci-Hub for the paper.
-        download_directory (str): Directory to save downloaded PDFs.
-        filename (str): Filename for the downloaded PDF.
-        is_oa (bool): Indicates if the paper is open access.
-        on_scihub (bool): Indicates if the paper is available on Sci-Hub.
-        pdf_urls (list): List of URLs to the PDF versions of the paper.
-        is_downloaded (bool): Indicates if the paper has been downloaded.
-        filepath (str): Path to the downloaded PDF file.
-        override_previous_attempt (bool): Whether to override a previous download attempt.
-        user_agents (list): List of user agents for making requests.
-    Methods:
-        download(): Finds and downloads the paper using the DOI or PMID.
-        check_open_access(): Checks if the paper is open access using Unpaywall.
-        check_pubmed_central_access(): Checks if the paper is available in PubMed Central.
-        check_crossref_access(doi): Checks if the paper is available on Crossref.
-        check_scihub_access(): Checks if the paper is available on Sci-Hub.
-        _download_pdf(): Downloads the PDF from the first available URL.
-        _create_json_sidecar(download_success, pdf_filepath, json_filepath, url=None): Creates a JSON sidecar file with download information.
-        _determine_paths(): Determines the file paths for the download.
-        _check_if_downloaded(download_directory_or_path, filetype=".pdf"): Checks if a non-corrupted file exists in the specified directory.
-        _look_for_previous_download(): Looks for a previous download of the PDF file.
-        _get_pdf_element(html_text, mirror): Extracts the PDF link from the HTML text of a Sci-Hub page.
+        doi (str): DOI encoded for safe file paths.
+        pmid (str | None): PubMed ID of the paper.
+        pdf_urls (list[str]): Candidate URLs pointing to PDF files.
+        filepath (str | None): Path to the downloaded PDF if successful.
+        is_downloaded (bool): ``True`` if the PDF has been retrieved.
+        is_oa (bool): ``True`` if the paper is open access.
+        on_scihub (bool): ``True`` if the PDF was found on Sci-Hub.
     """
 
[…]
