+++ pypaperretriever/reference_retriever.py
@@ -1,15 +1,24 @@
+"""Helpers to retrieve references and citations for scientific papers."""
+
+from __future__ import annotations
+
+import re
+from typing import Any, Dict, List, Optional
+
 import requests
 from Bio import Entrez
-import re
-from typing import List, Dict, Any, Optional
 
 from .utils import doi_to_pmid
 
+
 class ReferenceRetriever:
-    """
-    A class to retrieve academic paper references and citations using DOI or PMID identifiers.
-    
-    Fetches data from multiple sources including PubMed, Europe PMC, and CrossRef.
+    """Retrieve references and citing articles using a DOI or PMID.
+
+    Args:
+        email (str): Email address for API access.
+        doi (str | None): Digital Object Identifier.
+        pmid (str | None): PubMed identifier.
+        standardize (bool): If ``True`` the output dictionaries share common keys.
     """
     
     def __init__(
@@ -18,7 +27,7 @@ class ReferenceRetriever:
         doi: Optional[str] = None,
         pmid: Optional[str] = None,
         standardize: bool = True,
-    ) -> None:
+    ):
         """
         Initialize the retriever with either DOI or PMID.
 
@@ -40,11 +49,10 @@ class ReferenceRetriever:
             print(f"[ReferenceRetriever] Converted DOI {self.doi} to PMID: {self.pmid}")
 
     def fetch_references(self) -> List[Dict[str, Any]]:
-        """
-        Fetch references for the given DOI or PMID using multiple sources.
+        """Fetch references for the current paper.
 
         Returns:
-            List[Dict[str, Any]]: List of dictionaries containing reference metadata. Empty list if none found.
+            list[dict[str, Any]]: Reference metadata.
 
         Raises:
             ValueError: If neither DOI nor PMID is provided.
@@ -61,11 +69,10 @@ class ReferenceRetriever:
         return references
 
     def fetch_cited_by(self) -> List[Dict[str, Any]]:
-        """
-        Fetch articles that cite the given DOI or PMID.
+        """Fetch articles that cite the current paper.
 
         Returns:
-            List[Dict[str, Any]]: List of dictionaries containing citing article metadata. Empty list if none found.
+            list[dict[str, Any]]: Citing article metadata.
 
         Raises:
             ValueError: If PMID conversion fails or is not provided.
@@ -89,11 +96,10 @@ class ReferenceRetriever:
         return cited_by
 
     def get_paper_metadata(self) -> Dict[str, Any]:
-        """
-        Fetch metadata for the paper identified by DOI or PMID.
+        """Fetch metadata for the current paper.
 
         Returns:
-            Dict[str, Any]: Paper metadata including DOI, PMID, title, authors, and year. Empty dict if not found.
+            dict[str, Any]: Metadata including identifiers, title and authors.
         """
         print(f"[ReferenceRetriever] Fetching metadata for DOI: {self.doi}, PMID: {self.pmid}")
         if self.pmid:
@@ -120,12 +126,7 @@ class ReferenceRetriever:
         return {}
 
     def _find_references(self) -> List[Dict[str, Any]]:
-        """
-        Find references using multiple sources (PubMed, Europe PMC, CrossRef).
-
-        Returns:
-            List[Dict[str, Any]]: Combined list of references from all sources.
-        """
+        """Gather references from multiple sources."""
         print("[ReferenceRetriever] Finding references using multiple sources.")
         references = []
         if self.pmid:
@@ -154,12 +155,7 @@ class ReferenceRetriever:
         return references
 
     def _find_cited_by(self) -> List[Dict[str, Any]]:
-        """
-        Find citing articles using multiple sources (PubMed, Europe PMC).
-
-        Returns:
-            List[Dict[str, Any]]: Combined list of citing articles from all sources.
-        """
[…]
