+++ pypaperretriever/__init__.py
@@ -1,8 +1,18 @@
+"""Convenience imports for the public pypaperretriever API."""
+
+from __future__ import annotations
+
 from .paper_retriever import PaperRetriever
 from .pubmed_searcher import PubMedSearcher
 from .reference_retriever import ReferenceRetriever
 from .image_extractor import ImageExtractor
 from .paper_tracker import PaperTracker
-from .utils import encode_doi, decode_doi, pmid_to_doi, doi_to_pmid
+from .utils import decode_doi, doi_to_pmid, encode_doi, pmid_to_doi
 
-__all__ = ['PaperRetriever', 'PubMedSearcher', 'ReferenceRetriever', 'ImageExtractor', 'PaperTracker']
+__all__ = [
+    "PaperRetriever",
+    "PubMedSearcher",
+    "ReferenceRetriever",
+    "ImageExtractor",
+    "PaperTracker",
+]
