+++ .github/workflows/docs.yml
@@ -21,7 +21,7 @@ jobs:
         with:
           python-version: '3.x'
       - name: Install deps
-        run: pip install mkdocs mkdocs-material mkdocstrings[python]
+        run: pip install -r requirements-dev.txt
       - name: Build site
         run: mkdocs build --strict
       - name: Upload artifact

[…]
