### 2025-08-10 cb0a3c2
Revert "Merge pull request #6 from JosephIsaacTurner/codex/improve-docstrings-for-python-files"
This reverts commit ea55714a1726dfe339995162309e5c694a05a51c, reversing
changes made to ef472c0b8500bfff478acf648c728bbd14279546.
