### 2025-08-10 a4ce9b2
fix docs requirements
