+++ docs/requirements.txt
@@ -1,0 +1,4 @@
+mkdocs>=1.6
+mkdocs-material>=9.5
+mkdocstrings[python]>=0.25
+black>=24.0
