+++ .github/workflows/docs.yml
@@ -1,4 +1,5 @@
 name: docs
+
 on:
   push:
     branches: [main]
@@ -17,17 +18,24 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+
       - uses: actions/setup-python@v4
         with:
-          python-version: '3.x'
-      - name: Install deps
-        run: pip install -r requirements-dev.txt
+          python-version: '3.12'  # stable wheels for mkdocs deps
+
+      - name: Install docs deps
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r docs/requirements.txt
+
       - name: Build site
         run: mkdocs build --strict
+
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
           path: site
+
   deploy:
     needs: build
     runs-on: ubuntu-latest
