### 2025-08-10 41ea858
add legal/ethical note to README.md
