### 2025-08-31 a7c52da
Revise documentation for legal and ethical clarity
Updated the documentation to clarify the legal and ethical use of sources, emphasizing lawfully available PDFs and removing references to unauthorized sources. Enhanced descriptions of features and methods for improved clarity.
