+++ paper/paper.md
@@ -25,17 +25,19 @@ bibliography: paper.bib
 ---
 # Summary
 
-PyPaperRetriever is a Python tool that automates the discovery and retrieval of scientific literature, starting from either a DOI or PubMed ID. It queries multiple sources—Unpaywall, NIH Entrez/PMC, CrossRef, and optionally Sci-Hub—to locate and download PDFs, prioritizing open access. The tool supports programmatic PubMed searches with custom queries, enabling bulk PDF downloads, metadata resolution, and optional figure extraction in a single, reproducible workflow. Usable via command line or as a Python module, PyPaperRetriever provides a scalable, PubMed-first solution for literature reviews, dataset creation, and other projects requiring comprehensive literature aggregation.
+PyPaperRetriever is a Python tool that automates the discovery and retrieval of scientific literature, starting from either a DOI or PubMed ID. It queries multiple sources—Unpaywall, NIH Entrez/PMC, and Crossref—to locate and download lawfully available PDFs, prioritizing open access. The tool supports programmatic PubMed searches with custom queries, enabling at‑scale retrieval where lawfully available, metadata resolution, and optional figure extraction in a single, reproducible workflow. Usable via command line or as a Python module, PyPaperRetriever provides a scalable, PubMed‑first solution for literature reviews, dataset creation, and other projects requiring comprehensive literature aggregation.
+
+**Legal/Ethical Notice:** This article documents software that prioritizes lawfully available sources (e.g., PubMed Central, Unpaywall‑indexed OA, and publisher‑permitted copies). It does not instruct readers to access or download materials from unauthorized sources. Any references to third‑party websites that may host unauthorized copies are provided only for contextual completeness and are not endorsed. Readers are solely responsible for compliance with applicable laws and institutional policies.
 
 ### Core Features
 
 ![PyPaperRetriever paper retrieval workflow.\label{fig:workflow}](figure1.png)
 
-**1. Find and Download PDFs for a Given DOI or PubMed ID:** PyPaperRetriever can find and download PDFs for a given DOI or PubMed ID by querying multiple APIs, including Unpaywall, NIH's Entrez, CrossRef, and Sci-Hub. If a PDF is found, PyPaperRetriever will download it to the user's system. Sci-Hub is used as a last resort due to its controversial nature, and users can easily disable this feature if desired.
+**1. Find and Download PDFs for a Given DOI or PubMed ID:** PyPaperRetriever can find and download PDFs for a given DOI or PubMed ID by querying multiple APIs, including Unpaywall, NIH's Entrez, and Crossref. If a PDF is lawfully available (e.g., via PubMed Central or publisher‑permitted OA), PyPaperRetriever will download it to the user's system.
 
 ![PubMedSearcher workflow for searching PubMed programmatically.\label{fig:pubmed}](figure2.png)
 
-**2. Search PubMed programatically:** PyPaperRetriever can search PubMed using a query string and download PDFs of the search results. This feature is useful for researchers who want to download multiple papers based on a specific topic or keyword.
+**2. Search PubMed programmatically:** PyPaperRetriever can search PubMed using a query string and download PDFs of the search results where lawfully available (e.g., PubMed Central or publisher‑provided OA). This feature is useful for researchers who want to download multiple papers based on a specific topic or keyword.
 
 **3. Extract figures/images from PDFs:** PyPaperRetriever can extract figures and images from PDFs and save them as PNG files. This feature is useful for researchers who want to extract and analyze figures or images from scientific papers, and is robust to a variety of PDF formats. PyPaperRetriever has been used to build an extensive catalog of brain lesion images for training computer vision models.
 
@@ -43,11 +45,11 @@ PyPaperRetriever is a Python tool that automates the discovery and retrieval of 
 
 # Statement of Need
 
[…]
