### 2025-08-31 #8135 @arfon
@JosephIsaacTurner – As JOSS moves forward to publish this paper, I've proposed some small but important edits to the paper to keep the focus squarely on lawful access and to avoid any implication of encouraging infringement.

The Summary and feature descriptions now consistently emphasize retrieval only where lawfully available and highlight OA-first sources such as PubMed Central, Unpaywall-indexed OA, and publisher-permitted copies. The Legal/Ethical Notice clarifies that the article does not provide instructions for or endorse third-party sites that may host unauthorized copies.

While I understand the codebase includes an optional flag that can route requests via Sci-Hub, this capability is disabled by default, jurisdiction-dependent, and not documented or encouraged in the paper.

Hopefully you find these edits acceptable: https://github.com/JosephIsaacTurner/pypaperretriever/pull/8
