### 2025-09-06 #8135 @arfon
@xavieryao – I realize I never responded to [your comment](https://github.com/openjournals/joss-reviews/issues/8135#issuecomment-2920715684) back in May.

> Therefore, I recommend a major revision to address the need for this software and whether it meets "substantial scholarly effort".
>
> That being said, the software package and paper themselves are well-built. The package follows all guidelines for well-structured high-quality open source software. The only thing missing is a guideline for contributors.

Thanks for your feedback on this. I appreciate the points you’ve raised about overlap with existing tools and whether this represents a good/significant contribution. From my own inspection, I think this package is nicely put together, and unifies several capabilities, so I think it will be useful to the community.

In addition, while it’s on the smaller side of what we typically published, on balance I’m satisfied it meets our bar and I'm moving forward with acceptance.
