### 2025-09-06 b2d8403
Change affiliation format from list to string
