### 2025-09-06 #8135 @arfon
@JosephIsaacTurner – this PR will fix your paper: https://github.com/JosephIsaacTurner/pypaperretriever/pull/9
