### 2025-10-19 2ba2d65
update citing info
